+++ scripts/Main.js
@@ -1,11 +1,11 @@
 import { ItemAnimator } from "./animation/ItemAnimator.js";
-import { ObjectsInteractionsFX } from "./ObjectsInteractionsFX.js";
-import { ObjectsInteractionsFXData} from "./data/ObjectsInteractionsFXData.js";
+import { ObjectsInteractionsFX as OIF } from "./ObjectsInteractionsFX.js";
+import { ObjectsInteractionsFXData as OIFD } from "./data/ObjectsInteractionsFXData.js"
 import { ItemTags } from "./interface/ItemTags.js";
 
 Hooks.on("init", () =>
 {
-    ObjectsInteractionsFX.Initialize();
+    OIF.Initialize();
 
     Hooks.on("getItemSheetHeaderButtons", async (itemSheet, buttonArray) => {
         let TagButton = {
@@ -13,7 +13,7 @@ Hooks.on("init", () =>
             class: "item-tags",
             icon: "fas fa-tags",
             onclick: async () => {
-                await ObjectsInteractionsFXData.CreateData(itemSheet.object, [ '' ]);
+                await OIFD.CreateData(itemSheet.object, [ '' ]);
                 new ItemTags().render(true, { item: itemSheet.object, width: 480 })
             }
         }
@@ -21,7 +21,7 @@ Hooks.on("init", () =>
         buttonArray.unshift(TagButton);
     });
 
-    window[ObjectsInteractionsFX.FLAGS.OIF] = {
+    window[OIF.FLAGS.OIF] = {
         MeleeSingleAttack: ItemAnimator.MeleeSingleAttack,
         RangedSingleAttack: ItemAnimator.RangedSingleAttack
     };
@@ -34,6 +34,6 @@ Hooks.on("init", () =>
         padding:15pt;
     `);
 
-    Hooks.callAll("oifReady", game.modules.get(ObjectsInteractionsFX.ID).api);
+    Hooks.callAll("oifReady", game.modules.get(OIF.ID).api);
     console.log("Automated Objects, Interactions and Effects is ready!!");
 });