+++ scripts/animation/ItemAnimator.js
@@ -15,8 +15,14 @@ export class ItemAnimator
         // Calculate distance
         let Distance = canvas.grid.measureDistance(author, target);
 
-        // Is the Item a throwable?
-        if (item.data.data.properties.thr == true && Distance >= 3.0)
+        // Caculate the melee attack distance of the item
+        let WeaponMeleeDistance = canvas.dimensions.distance;
+        if (item.data.data.properties.rch) {
+            WeaponMeleeDistance += WeaponMeleeDistance;
+        }
+
+        // Is the Item a throwable and outside of melee range?
+        if (item.data.data.properties.thr == true && Distance >= WeaponMeleeDistance + canvas.dimensions.distance)
         {
             // Check if item should be removed
             if (game.settings.get(OIF.ID, "removeThrowableItem"))
@@ -29,21 +35,30 @@ export class ItemAnimator
                 }
             }
 
+            // Call hook
+            Hooks.call(OIF.HOOKS.WEAPON.MELEE.THROW.PRE, options);
+
             // Define throw sequence to be played
             let SequencerHelper = `${options.name}-throw-${author.data._id}`;
             let SequencerEffect = new Sequence()
                 .effect()
-                    .file(options.throwAnimation)
+                    .file(options.throwAnimation.effect)
                     .atLocation(author)
                     .stretchTo(target)
                     .name(SequencerHelper)
                     .missed(options.miss)
+                    .waitUntilFinished(-600)
             
             // Play throw sequence
             await SequencerEffect.play();
 
             // Get results from played sequence
             let [Effect] = Sequencer.EffectManager.getEffects({ name: SequencerHelper });
+            options.landedPosX = Effect.targetPosition.x - canvas.dimensions.size / 2;
+            options.landedPosY = Effect.targetPosition.y - canvas.dimensions.size / 2;
+
+            // Call hook
+            Hooks.call(OIF.HOOKS.WEAPON.MELEE.THROW.POS, options);
 
             // Create a copy of the item
             let ItemCopy = item.toObject();
@@ -53,8 +68,8 @@ export class ItemAnimator
             if (options.miss && OIF.SETTINGS.LOADED_MODULES.ITEM_PILES && game.settings.get(OIF.ID, "createItemPilesOnMiss")) {
                 // Get the position where the item landed
                 let ItemPilePosition = {
-                    x: Effect.targetPosition.x - 50,
-                    y: Effect.targetPosition.y - 50
+                    x: options.landedPosX,
+                    y: options.landedPosY,
                 }
 
                 // Drop item
@@ -68,20 +83,27 @@ export class ItemAnimator
         }
         else
         {
+            // Call hook
+            Hooks.call(OIF.HOOKS.WEAPON.MELEE.HIT.PRE, options);
+
             // Define melee sequence to be played
             let SequencerHelper = `${options.name}-melee-${author.data._id}`;
             let SequencerEffect = new Sequence()
                 .effect()
-                    .file(options.meleeAnimation)
+                    .file(options.meleeAnimation.effect)
                     .atLocation(author)
                     .stretchTo(target)
                     .name(SequencerHelper)
-                    .waitUntilFinished(options.powerfulDelay ?? 0)
+                    .waitUntilFinished(options.meleeAnimation.powerfulDelay ?? 0)
             
             // Play melee sequence
             await SequencerEffect.play();
+
+            // Call hook
+            Hooks.call(OIF.HOOKS.WEAPON.MELEE.HIT.POS, options);
+
             // Check if impact effect should be played
-            if (options.powerful == true && game.settings.get(OIF.ID, "powerfulImpactShakeEffect")) {
+            if (options.meleeAnimation.powerful == true && game.settings.get(OIF.ID, "powerfulImpactShakeEffect")) {
                 KFC.executeForEveryone("earthquake", 1, 500, 1)
             }
         }
@@ -107,41 +129,63 @@ export class ItemAnimator
             }
             else
             {
+                // Create a copy of the ammunition item
+                options.ammunitionItem = await author.actor.getEmbeddedDocument("Item", item.data.data.consume.target);
+
+                // Call hook
+                Hooks.call(OIF.HOOKS.WEAPON.RANGED.HIT.PRE, options);
+
                 // Define ranged sequence to be played
                 let SequencerHelper = `${options.name}-ranged-${author.data._id}`;
                 let SequencerEffect = new Sequence()
                     .effect()
-                        .file(options.rangedAnimation)
+                        .file(options.rangedAnimation.effect)
                         .atLocation(author)
                         .stretchTo(target)
                         .name(SequencerHelper)
                         .missed(options.miss)
+                        .waitUntilFinished(-1150)
 
                 // Play ranged sequence
                 await SequencerEffect.play();
 
+                //// Call hook
+                //Hooks.call(OIF.HOOKS.WEAPON.RANGED.HIT.SOUND.PRE);
+                //
+                //// Define sound sequence to be played
+                //let SequencerSound = new Sequence()
+                //    .sound()
+                //        .file(option)
+                //        .volume(1.0)
+                //        .startTime(500)
+                //
+                //// Play sound sequence
+                //SequencerSound.play();
+
                 // Get results from played sequence
                 let [Effect] = Sequencer.EffectManager.getEffects({ name: SequencerHelper });
+                options.landedPosX = Effect.targetPosition.x - canvas.dimensions.size / 2;
+                options.landedPosY = Effect.targetPosition.y - canvas.dimensions.size / 2;
 
-                // Create a copy of the ammunition item
-                let Ammunition = await author.actor.getEmbeddedDocument("Item", item.data.data.consume.target);
+                // Call hook
+                Hooks.call(OIF.HOOKS.WEAPON.RANGED.HIT.POS, options);
 
                 // Check if Item Pile should be created
                 if (options.miss && OIF.SETTINGS.LOADED_MODULES.ITEM_PILES && game.settings.get(OIF.ID, "createItemPilesOnMiss"))
                 {
                     // Get the position where the item landed
                     let ItemPilePosition = {
-                        x: Effect.targetPosition.x - 50,
-                        y: Effect.targetPosition.y - 50
+                        x: options.landedPosX,
+                        y: options.landedPosY,
                     }
 
                     // Drop item
-                    ItemDropper.DropAt(Ammunition, 1, ItemPilePosition, target.data.elevation)
+                    ItemDropper.DropAt(options.ammunitionItem, 1, ItemPilePosition, target.data.elevation)
                 }
                 else if (game.settings.get(OIF.ID, "addAmmunitionToTargetInventory"))
                 {
                     // Add the ammunition to Target's inventory
-                    InventoryManipulator.AddItem(target, Ammunition, 1);
+                    InventoryManipulator.AddItem(target, options.ammunitionItem, 1);
                 }
             }
         }
@@ -151,7 +195,7 @@ export class ItemAnimator
             let SequencerHelper = `${options.name}-ranged-${author.data._id}`;
             let SequencerEffect = new Sequence()
                 .effect()
-                    .file(options.rangedAnimation)
+                    .file(options.rangedAnimation.effect)
                     .atLocation(author)
                     .stretchTo(target)
                     .name(SequencerHelper)